+++ tests/test_generator.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytest
 
 from zhinst.toolkit import Waveforms
-
+from zhinst.core import compile_seqc
 
 @pytest.fixture()
 def generator(data_dir, mock_connection, shfqa):
@@ -19,33 +19,26 @@ def generator(data_dir, mock_connection, shfqa):
 
     yield shfqa.qachannels[0].generator
 
+def test_enable_sequencer(shfqa, mock_connection):
+    shfqa.qachannels[0].generator.enable_sequencer(single=True)
+    mock_connection.return_value.set.assert_called_with(
+        "/dev1234/qachannels/0/generator/single", 1
+    )
+    mock_connection.return_value.syncSetInt.assert_called_with(
+        "/dev1234/qachannels/0/generator/enable", 1
+    )
 
-def test_enable_sequencer(generator, mock_connection):
-    with patch("zhinst.toolkit.driver.nodes.generator.utils", autospec=True) as utils:
-        generator.enable_sequencer(single=True)
-        utils.enable_sequencer.assert_called_once_with(
-            mock_connection.return_value,
-            "DEV1234",
-            0,
-            single=1,
-        )
-        generator.enable_sequencer(single=False)
-        utils.enable_sequencer.assert_called_with(
-            mock_connection.return_value,
-            "DEV1234",
-            0,
-            single=0,
-        )
+    yield shfqa.qachannels[0].generator
 
 
 def test_load_sequencer_program_empty_string_none(generator):
-    with pytest.raises(ValueError):
+    with pytest.raises(RuntimeError):
         generator.load_sequencer_program("")
-    with pytest.raises(ValueError):
+    with pytest.raises(RuntimeError):
         generator.load_sequencer_program(None)
 
 
-def test_wait_done(mock_connection, generator):
+def test_wait_done(mock_connection,shfqa, generator):
     single = 0
     enable = iter([])
 
@@ -59,45 +52,39 @@ def test_wait_done(mock_connection, generator):
 
     # if not single mode this function throws a RuntimeError
     with pytest.raises(RuntimeError) as e_info:
-        generator.wait_done()
+        shfqa.qachannels[0].generator.wait_done()
 
     # already finished
     single = 1
     enable = iter([0] * 2)
-    generator.wait_done()
+    shfqa.qachannels[0].generator.wait_done()
     # finishes in time
     single = 1
     enable = iter([1] * 3 + [0] * 2)
-    generator.wait_done()
+    shfqa.qachannels[0].generator.wait_done()
     # don't finish
     single = 1
     enable = cycle([1])
     with pytest.raises(TimeoutError) as e_info:
-        generator.wait_done(timeout=0.1)
+        shfqa.qachannels[0].generator.wait_done(timeout=0.1)
 
 
-def test_load_sequencer_program(session, generator, mock_connection):
-    with patch("zhinst.toolkit.driver.nodes.generator.utils", autospec=True) as utils:
-        generator.load_sequencer_program("Test")
-        utils.load_sequencer_program.assert_called_once_with(
-            mock_connection.return_value,
-            "DEV1234",
-            0,
-            "Test",
-            timeout=10,
-        )
+def test_load_sequencer_program(shfqa, generator, mock_connection):
+    # everything ok
+    elf, info_original = compile_seqc("setTrigger(1);", "SHFQA4", [], 0)
+    info = shfqa.qachannels[0].generator.load_sequencer_program("setTrigger(1);")
+    mock_connection.return_value.set.assert_called_once()
+    assert (
+        mock_connection.return_value.set.call_args[0][0]
+        == "/dev1234/qachannels/0/generator/elf/data"
+    )
+    assert all(
+        mock_connection.return_value.set.call_args[0][1]
+        == np.frombuffer(elf, dtype="uint32")
+    )
+    assert info == info_original
 
-        generator.load_sequencer_program("Test", timeout=1)
-        utils.load_sequencer_program.assert_called_with(
-            mock_connection.return_value,
-            "DEV1234",
-            0,
-            "Test",
-            timeout=1,
-        )
-
-
-def test_write_to_waveform_memory(generator, mock_connection):
+def test_write_to_waveform_memory(shfqa, generator, mock_connection):
 
     waveforms = Waveforms()
     waveforms[0] = np.ones(1000)
@@ -106,40 +93,46 @@ def test_write_to_waveform_memory(generator, mock_connection):
 
     waveforms_long = Waveforms()
     waveforms_long[1000] = np.zeros(1000)
-    with patch("zhinst.toolkit.driver.nodes.generator.utils", autospec=True) as utils:
-        generator.write_to_waveform_memory(waveforms)
-        complex_wave = np.empty(1000, dtype=np.complex128)
-        complex_wave.real = np.ones(1000)
-        complex_wave.imag = np.ones(1000)
-        complex_wave = complex_wave
-        assert (
-            utils.write_to_waveform_memory.call_args[0][0]
-            == mock_connection.return_value
-        )
-        assert utils.write_to_waveform_memory.call_args[0][1] == "DEV1234"
-        assert utils.write_to_waveform_memory.call_args[0][2] == 0
-        assert np.allclose(
-            utils.write_to_waveform_memory.call_args[0][3][1],
-            np.ones(1000, dtype=np.complex128),
-        )
-        assert np.allclose(
-            utils.write_to_waveform_memory.call_args[0][3][2], complex_wave
-        )
-        assert utils.write_to_waveform_memory.call_args[1]["clear_existing"] == True
 
-        generator.write_to_waveform_memory(waveforms, clear_existing=False)
-        assert utils.write_to_waveform_memory.call_args[1]["clear_existing"] == False
+    generator.write_to_waveform_memory(waveforms)
+    complex_wave = np.empty(1000, dtype=np.complex128)
+    complex_wave.real = np.ones(1000)
+    complex_wave.imag = np.ones(1000)
+    complex_wave = complex_wave
 
-        generator.write_to_waveform_memory({0: np.ones(1000)}, clear_existing=True)
-        assert all(utils.write_to_waveform_memory.call_args[0][3][0] == np.ones(1000))
+    assert mock_connection.return_value.set.call_args[0][0][0] == ('/dev1234/qachannels/0/generator/clearwave', 1)
+    assert mock_connection.return_value.set.call_args[0][0][1][0] == '/dev1234/qachannels/0/generator/waveforms/0/wave'
+    assert all (mock_connection.return_value.set.call_args[0][0][1][1] == np.ones(1000, dtype=np.complex128))
+    assert mock_connection.return_value.set.call_args[0][0][2][0] == '/dev1234/qachannels/0/generator/waveforms/1/wave'
+    assert all (mock_connection.return_value.set.call_args[0][0][2][1] == np.ones(1000, dtype=np.complex128))
+    assert mock_connection.return_value.set.call_args[0][0][3][0] == '/dev1234/qachannels/0/generator/waveforms/2/wave'
+    assert all (mock_connection.return_value.set.call_args[0][0][3][1] == complex_wave)
+
+    shfqa.qachannels[0].generator.write_to_waveform_memory(
+        waveforms, clear_existing=False
+    )
+    assert mock_connection.return_value.set.call_args[0][0][0][0] == '/dev1234/qachannels/0/generator/waveforms/0/wave'
+    assert all (mock_connection.return_value.set.call_args[0][0][0][1] == np.ones(1000, dtype=np.complex128))
+    assert mock_connection.return_value.set.call_args[0][0][1][0] == '/dev1234/qachannels/0/generator/waveforms/1/wave'
+    assert all (mock_connection.return_value.set.call_args[0][0][1][1] == np.ones(1000, dtype=np.complex128))
+    assert mock_connection.return_value.set.call_args[0][0][2][0] == '/dev1234/qachannels/0/generator/waveforms/2/wave'
+    assert all (mock_connection.return_value.set.call_args[0][0][2][1] == complex_wave)
+
+    shfqa.qachannels[0].generator.write_to_waveform_memory(
+        {0: np.ones(1000)}, clear_existing=True
+    )
+    assert mock_connection.return_value.set.call_args[0][0][0] == ('/dev1234/qachannels/0/generator/clearwave', 1)
+    assert mock_connection.return_value.set.call_args[0][0][1][0] == '/dev1234/qachannels/0/generator/waveforms/0/wave'
+    assert all (mock_connection.return_value.set.call_args[0][0][1][1] == np.ones(1000, dtype=np.complex128))
+
 
     with pytest.raises(RuntimeError) as e_info:
-        generator.write_to_waveform_memory(waveforms_long)
+        shfqa.qachannels[0].generator.write_to_waveform_memory(waveforms_long)
 
 
-def test_read_from_waveform_memory(generator, mock_connection):
+def test_read_from_waveform_memory(shfqa, mock_connection):
 
-    result = generator.read_from_waveform_memory()
+    result = shfqa.qachannels[0].generator.read_from_waveform_memory()
     assert len(result) == 0
     mock_connection.return_value.get.assert_called_once_with(
         "/dev1234/qachannels/0/generator/waveforms/*/wave",
@@ -147,7 +140,7 @@ def test_read_from_waveform_memory(generator, mock_connection):
         flat=True,
     )
 
-    result = generator.read_from_waveform_memory([0, 3])
+    result = shfqa.qachannels[0].generator.read_from_waveform_memory([0, 3])
     mock_connection.return_value.get.assert_called_with(
         "/dev1234/qachannels/0/generator/waveforms/0/wave,/dev1234/qachannels/0/generator/waveforms/3/wave",
         settingsonly=False,
@@ -162,7 +155,7 @@ def test_read_from_waveform_memory(generator, mock_connection):
             {"vector": np.ones(1000, dtype=np.complex128)}
         ],
     }
-    result = generator.read_from_waveform_memory()
+    result = shfqa.qachannels[0].generator.read_from_waveform_memory()
     assert len(result) == 2
     np.allclose(result[0][0], np.ones(1000, dtype=np.complex128))
     assert not result[0][1]
@@ -171,7 +164,7 @@ def test_read_from_waveform_memory(generator, mock_connection):
 
 
 def test_configure_sequencer_triggering(generator, mock_connection):
-    with patch("zhinst.toolkit.driver.nodes.generator.utils", autospec=True) as utils:
+    with patch("zhinst.toolkit.driver.devices.shfqa.utils", autospec=True) as utils:
         generator.configure_sequencer_triggering(
             aux_trigger="fobarob", play_pulse_delay=0.0001
         )
@@ -184,8 +177,8 @@ def test_configure_sequencer_triggering(generator, mock_connection):
         )
 
 
-def test_available_aux_trigger_inputs(generator):
-    assert generator.available_aux_trigger_inputs == [
+def test_available_aux_trigger_inputs(shfqa):
+    assert shfqa.qachannels[0].generator.available_aux_trigger_inputs == [
         "chan0trigin0",
         "chan0trigin1",
         "chan1trigin0",