+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   devices connected to the data server. (`#134`)
 * Bugfix: Nodes with nameless options don't raise an exception when their enum attribute is called (`#165`).
 * Bugfix: Values of enumerated nodes can now be pickled (`#129`).
+* Adapt toolkit to use the offline awg compiler when uploading a sequencer code to
+  a awg node. Improves the performance a lot and also enables the uploading of
+  a sequencer code within a transaction (Works both for AWGs and Generators).
+* Add new function `compile_sequencer_code` to the awg node.
 
 ## Version 0.3.5
 * Adapt AWG Waveform upload (`write_to_waveform_memory`) to append to existing transactions.